+++ src/Cli.Tests/AddEntityTests.cs
@@ -105,7 +105,7 @@ namespace Cli.Tests
                 );
 
             string initialConfiguration = AddPropertiesToJson(INITIAL_CONFIG, GetFirstEntityConfiguration());
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(initialConfiguration, out RuntimeConfig? runtimeConfig), "Loaded config");
+            Assert.IsTrue(FileSystemRuntimeConfigLoader.TryParseConfig(initialConfiguration, out RuntimeConfig? runtimeConfig), "Loaded config");
 
             Assert.IsFalse(TryAddNewEntity(options, runtimeConfig, out RuntimeConfig updatedRuntimeConfig));
 
@@ -279,7 +279,7 @@ namespace Cli.Tests
                 graphQLOperationForStoredProcedure: null
                 );
 
-            RuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
 
             Assert.AreEqual(expectSuccess, TryAddNewEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -369,7 +369,7 @@ namespace Cli.Tests
                 graphQLOperationForStoredProcedure: graphQLOperation
             );
 
-            RuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryAddNewEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -405,7 +405,7 @@ namespace Cli.Tests
                 graphQLOperationForStoredProcedure: null
                 );
 
-            RuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(INITIAL_CONFIG, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryAddNewEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -430,7 +430,7 @@ namespace Cli.Tests
 
         private Task ExecuteVerifyTest(AddOptions options, string config = INITIAL_CONFIG, VerifySettings? settings = null)
         {
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(config, out RuntimeConfig? runtimeConfig), "Loaded base config.");
+            Assert.IsTrue(FileSystemRuntimeConfigLoader.TryParseConfig(config, out RuntimeConfig? runtimeConfig), "Loaded base config.");
 
             Assert.IsTrue(TryAddNewEntity(options, runtimeConfig, out RuntimeConfig updatedRuntimeConfig), "Added entity to config.");
