+++ src/Cli.Tests/EndToEndTests.cs
@@ -16,7 +16,7 @@ public class EndToEndTests
     : VerifyBase
 {
     private IFileSystem? _fileSystem;
-    private RuntimeConfigLoader? _runtimeConfigLoader;
+    private FileSystemRuntimeConfigLoader? _runtimeConfigLoader;
     private ILogger<Program>? _cliLogger;
 
     [TestInitialize]
@@ -29,7 +29,7 @@ public class EndToEndTests
 
         _fileSystem = fileSystem;
 
-        _runtimeConfigLoader = new RuntimeConfigLoader(_fileSystem);
+        _runtimeConfigLoader = new FileSystemRuntimeConfigLoader(_fileSystem);
 
         ILoggerFactory loggerFactory = TestLoggerSupport.ProvisionLoggerFactory();
