+++ src/Cli.Tests/UpdateEntityTests.cs
@@ -524,7 +524,7 @@ namespace Cli.Tests
                 sourceKeyFields: keyFields);
 
             string initialConfig = AddPropertiesToJson(INITIAL_CONFIG, initialSourceObjectEntity);
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
             Assert.AreEqual(expectSuccess, TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig updatedConfig));
 
             if (expectSuccess)
@@ -580,7 +580,7 @@ namespace Cli.Tests
                 sourceType: "stored-procedure");
 
             string initialConfig = AddPropertiesToJson(INITIAL_CONFIG, SINGLE_ENTITY_WITH_STORED_PROCEDURE);
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.AreEqual(isSuccess, TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -646,7 +646,7 @@ namespace Cli.Tests
                         }
                     }";
 
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -781,7 +781,7 @@ namespace Cli.Tests
                 graphQLOperationForStoredProcedure: graphQLOperation);
 
             string initialConfig = AddPropertiesToJson(INITIAL_CONFIG, SP_DEFAULT_REST_METHODS_GRAPHQL_OPERATION);
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -817,7 +817,7 @@ namespace Cli.Tests
                         }
                     }";
 
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -853,7 +853,7 @@ namespace Cli.Tests
 
             string initialConfig = AddPropertiesToJson(INITIAL_CONFIG, SINGLE_ENTITY_WITH_STORED_PROCEDURE);
 
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -922,7 +922,7 @@ namespace Cli.Tests
                         }
                     }";
 
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -960,7 +960,7 @@ namespace Cli.Tests
                             }
                         }
                     }";
-            RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig);
 
             Assert.IsFalse(TryUpdateExistingEntity(options, runtimeConfig!, out RuntimeConfig _));
         }
@@ -1127,7 +1127,7 @@ namespace Cli.Tests
 
         private Task ExecuteVerifyTest(string initialConfig, UpdateOptions options, VerifySettings? settings = null)
         {
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig), "Parsed config file.");
+            Assert.IsTrue(FileSystemRuntimeConfigLoader.TryParseConfig(initialConfig, out RuntimeConfig? runtimeConfig), "Parsed config file.");
 
             Assert.IsTrue(TryUpdateExistingEntity(options, runtimeConfig, out RuntimeConfig updatedRuntimeConfig), "Successfully added entity to config.");
