+++ src/Cli.Tests/UtilsTests.cs
@@ -110,8 +110,8 @@ public class UtilsTests
     [DataTestMethod]
     [DataRow("", "my-config.json", "my-config.json", DisplayName = "user provided the config file and environment variable was not set.")]
     [DataRow("Test", "my-config.json", "my-config.json", DisplayName = "user provided the config file and environment variable was set.")]
-    [DataRow("Test", null, $"{RuntimeConfigLoader.CONFIGFILE_NAME}.Test{RuntimeConfigLoader.CONFIG_EXTENSION}", DisplayName = "config not provided, but environment variable was set.")]
-    [DataRow("", null, $"{RuntimeConfigLoader.CONFIGFILE_NAME}{RuntimeConfigLoader.CONFIG_EXTENSION}", DisplayName = "neither config was provided, nor environment variable was set.")]
+    [DataRow("Test", null, $"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.Test{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}", DisplayName = "config not provided, but environment variable was set.")]
+    [DataRow("", null, $"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}", DisplayName = "neither config was provided, nor environment variable was set.")]
     public void TestConfigSelectionBasedOnCliPrecedence(
         string? environmentValue,
         string? userProvidedConfigFile,
@@ -120,13 +120,13 @@ public class UtilsTests
         MockFileSystem fileSystem = new();
         fileSystem.AddFile(expectedRuntimeConfigFile, new MockFileData(""));
 
-        RuntimeConfigLoader loader = new(fileSystem);
+        FileSystemRuntimeConfigLoader loader = new(fileSystem);
 
-        string? envValueBeforeTest = Environment.GetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME);
-        Environment.SetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, environmentValue);
+        string? envValueBeforeTest = Environment.GetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME);
+        Environment.SetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, environmentValue);
         Assert.IsTrue(TryGetConfigFileBasedOnCliPrecedence(loader, userProvidedConfigFile, out string? actualRuntimeConfigFile));
         Assert.AreEqual(expectedRuntimeConfigFile, actualRuntimeConfigFile);
-        Environment.SetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, envValueBeforeTest);
+        Environment.SetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, envValueBeforeTest);
     }
 
     /// <summary>
@@ -242,12 +242,12 @@ public class UtilsTests
     public void TestMergeConfig()
     {
         MockFileSystem fileSystem = new();
-        fileSystem.AddFile(RuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(BASE_CONFIG));
+        fileSystem.AddFile(FileSystemRuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(BASE_CONFIG));
         fileSystem.AddFile("dab-config.Test.json", new MockFileData(ENV_BASED_CONFIG));
 
-        RuntimeConfigLoader loader = new(fileSystem);
+        FileSystemRuntimeConfigLoader loader = new(fileSystem);
 
-        Environment.SetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, "Test");
+        Environment.SetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, "Test");
 
         Assert.IsTrue(Cli.ConfigMerger.TryMergeConfigsIfAvailable(fileSystem, loader, new StringLogger(), out string? mergedConfig), "Failed to merge config files");
         Assert.AreEqual(mergedConfig, "dab-config.Test.merged.json");
@@ -285,7 +285,7 @@ public class UtilsTests
         MockFileSystem fileSystem = new();
 
         // Setting up the test scenarios
-        Environment.SetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, environmentValue);
+        Environment.SetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, environmentValue);
         string baseConfig = "dab-config.json";
         string envBasedConfig = "dab-config.Test.json";
 
@@ -299,7 +299,7 @@ public class UtilsTests
             fileSystem.AddFile(envBasedConfig, new("{}"));
         }
 
-        RuntimeConfigLoader loader = new(fileSystem);
+        FileSystemRuntimeConfigLoader loader = new(fileSystem);
 
         Assert.AreEqual(
             expectedIsMergedConfigAvailable,
@@ -307,7 +307,7 @@ public class UtilsTests
             "Availability of merge config should match");
         Assert.AreEqual(expectedMergedConfigFileName, mergedConfigFile, "Merge config file name should match expected");
 
-        Environment.SetEnvironmentVariable(RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, null);
+        Environment.SetEnvironmentVariable(FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, null);
     }
 }
