+++ src/Service.Tests/Authentication/Helpers/WebHostBuilderHelper.cs
@@ -41,7 +41,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Authentication.Helpers
         {
             // Setup RuntimeConfigProvider object for the pipeline.
             MockFileSystem fileSystem = new();
-            RuntimeConfigLoader loader = new(fileSystem);
+            FileSystemRuntimeConfigLoader loader = new(fileSystem);
             RuntimeConfigProvider runtimeConfigProvider = new(loader);
 
             return await new HostBuilder()
