+++ src/Service.Tests/Configuration/ConfigurationTests.cs
@@ -40,7 +40,7 @@ using Moq;
 using MySqlConnector;
 using Npgsql;
 using VerifyMSTest;
-using static Azure.DataApiBuilder.Config.RuntimeConfigLoader;
+using static Azure.DataApiBuilder.Config.FileSystemRuntimeConfigLoader;
 
 namespace Azure.DataApiBuilder.Service.Tests.Configuration
 {
@@ -202,7 +202,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
                 Assert.IsTrue(runtimeConfigJson.Contains(sourceTypeName));
             }
 
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(runtimeConfigJson, out RuntimeConfig deserializedRuntimeConfig));
+            Assert.IsTrue(FileSystemRuntimeConfigLoader.TryParseConfig(runtimeConfigJson, out RuntimeConfig deserializedRuntimeConfig));
 
             Assert.IsTrue(deserializedRuntimeConfig.Entities.ContainsKey("MyEntity"));
             Assert.AreEqual("sourceName", deserializedRuntimeConfig.Entities["MyEntity"].Source.Object);
@@ -571,7 +571,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         [TestMethod("Validates if deserialization of MsSql config file succeeds."), TestCategory(TestCategory.MSSQL)]
         public Task TestReadingRuntimeConfigForMsSql()
         {
-            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{RuntimeConfigLoader.CONFIGFILE_NAME}.{MSSQL_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}"));
+            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{MSSQL_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}"));
         }
 
         /// <summary>
@@ -581,7 +581,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         [TestMethod("Validates if deserialization of MySql config file succeeds."), TestCategory(TestCategory.MYSQL)]
         public Task TestReadingRuntimeConfigForMySql()
         {
-            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{RuntimeConfigLoader.CONFIGFILE_NAME}.{MYSQL_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}"));
+            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{MYSQL_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}"));
         }
 
         /// <summary>
@@ -591,7 +591,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         [TestMethod("Validates if deserialization of PostgreSql config file succeeds."), TestCategory(TestCategory.POSTGRESQL)]
         public Task TestReadingRuntimeConfigForPostgreSql()
         {
-            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{RuntimeConfigLoader.CONFIGFILE_NAME}.{POSTGRESQL_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}"));
+            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{POSTGRESQL_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}"));
         }
 
         /// <summary>
@@ -601,7 +601,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         [TestMethod("Validates if deserialization of the CosmosDB_NoSQL config file succeeds."), TestCategory(TestCategory.COSMOSDBNOSQL)]
         public Task TestReadingRuntimeConfigForCosmos()
         {
-            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{RuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}"));
+            return ConfigFileDeserializationValidationHelper(File.ReadAllText($"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}"));
         }
 
         /// <summary>
@@ -611,7 +611,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         /// <param name="runtimeConfig"></param>
         private Task ConfigFileDeserializationValidationHelper(string jsonString)
         {
-            Assert.IsTrue(RuntimeConfigLoader.TryParseConfig(jsonString, out RuntimeConfig runtimeConfig), "Deserialization of the config file failed.");
+            Assert.IsTrue(FileSystemRuntimeConfigLoader.TryParseConfig(jsonString, out RuntimeConfig runtimeConfig), "Deserialization of the config file failed.");
             return Verify(runtimeConfig);
         }
 
@@ -625,8 +625,8 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
             Environment.SetEnvironmentVariable(ASP_NET_CORE_ENVIRONMENT_VAR_NAME, MSSQL_ENVIRONMENT);
             string[] args = new[]
             {
-                $"--ConfigFileName={RuntimeConfigLoader.CONFIGFILE_NAME}." +
-                $"{COSMOS_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}"
+                $"--ConfigFileName={FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}." +
+                $"{COSMOS_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}"
             };
 
             TestServer server = new(Program.CreateWebHostBuilder(args));
@@ -644,7 +644,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
             Environment.SetEnvironmentVariable(
                 ASP_NET_CORE_ENVIRONMENT_VAR_NAME, MSSQL_ENVIRONMENT);
             Environment.SetEnvironmentVariable(
-                RuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, COSMOS_ENVIRONMENT);
+                FileSystemRuntimeConfigLoader.RUNTIME_ENVIRONMENT_VAR_NAME, COSMOS_ENVIRONMENT);
 
             TestServer server = new(Program.CreateWebHostBuilder(Array.Empty<string>()));
 
@@ -655,7 +655,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         public void TestConfigIsValid()
         {
             TestHelper.SetupDatabaseEnvironment(MSSQL_ENVIRONMENT);
-            RuntimeConfigLoader configPath = TestHelper.GetRuntimeConfigLoader();
+            FileSystemRuntimeConfigLoader configPath = TestHelper.GetRuntimeConfigLoader();
             RuntimeConfigProvider configProvider = TestHelper.GetRuntimeConfigProvider(configPath);
 
             Mock<ILogger<RuntimeConfigValidator>> configValidatorLogger = new();
@@ -675,26 +675,26 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         /// has highest precedence irrespective of what the connection string is in the config file.
         /// Verifying the Exception thrown.
         /// </summary>
-        [TestMethod($"Validates that environment variable {RuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING} has highest precedence."), TestCategory(TestCategory.COSMOSDBNOSQL)]
+        [TestMethod($"Validates that environment variable {FileSystemRuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING} has highest precedence."), TestCategory(TestCategory.COSMOSDBNOSQL)]
         public void TestConnectionStringEnvVarHasHighestPrecedence()
         {
             Environment.SetEnvironmentVariable(ASP_NET_CORE_ENVIRONMENT_VAR_NAME, COSMOS_ENVIRONMENT);
             Environment.SetEnvironmentVariable(
-                RuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING,
+                FileSystemRuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING,
                 "Invalid Connection String");
 
             try
             {
                 TestServer server = new(Program.CreateWebHostBuilder(Array.Empty<string>()));
                 _ = server.Services.GetService(typeof(CosmosClientProvider)) as CosmosClientProvider;
-                Assert.Fail($"{RuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING} is not given highest precedence");
+                Assert.Fail($"{FileSystemRuntimeConfigLoader.RUNTIME_ENV_CONNECTION_STRING} is not given highest precedence");
             }
             catch (Exception e)
             {
                 Assert.AreEqual(typeof(ApplicationException), e.GetType());
                 Assert.AreEqual(
                     $"Could not initialize the engine with the runtime config file: " +
-                    $"{RuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}",
+                    $"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}",
                     e.Message);
             }
         }
@@ -716,7 +716,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         {
             MockFileSystem fileSystem = new();
             fileSystem.AddFile(expectedRuntimeConfigFile, new MockFileData(string.Empty));
-            RuntimeConfigLoader runtimeConfigLoader = new(fileSystem);
+            FileSystemRuntimeConfigLoader runtimeConfigLoader = new(fileSystem);
 
             Environment.SetEnvironmentVariable(ASP_NET_CORE_ENVIRONMENT_VAR_NAME, hostingEnvironmentValue);
             Environment.SetEnvironmentVariable(RUNTIME_ENVIRONMENT_VAR_NAME, environmentValue);
@@ -764,7 +764,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
             const string CUSTOM_CONFIG = "custom-config.json";
             TestHelper.SetupDatabaseEnvironment(MSSQL_ENVIRONMENT);
             FileSystem fileSystem = new();
-            RuntimeConfigLoader loader = new(fileSystem);
+            FileSystemRuntimeConfigLoader loader = new(fileSystem);
             loader.TryLoadKnownConfig(out RuntimeConfig config);
 
             RuntimeConfig configWithCustomHostMode = config with
@@ -1087,7 +1087,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
             TestHelper.SetupDatabaseEnvironment(TestCategory.MSSQL);
 
             FileSystem fileSystem = new();
-            RuntimeConfigLoader loader = new(fileSystem);
+            FileSystemRuntimeConfigLoader loader = new(fileSystem);
 
             RuntimeConfigProvider configProvider = TestHelper.GetRuntimeConfigProvider(loader);
             RuntimeConfig config = configProvider.GetConfig();
@@ -1613,7 +1613,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
 
         private static ConfigurationPostParameters GetCosmosConfigurationParameters()
         {
-            string cosmosFile = $"{RuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}";
+            string cosmosFile = $"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}";
             return new(
                 File.ReadAllText(cosmosFile),
                 File.ReadAllText("schema.gql"),
@@ -1623,7 +1623,7 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
 
         private static ConfigurationPostParametersV2 GetCosmosConfigurationParametersV2()
         {
-            string cosmosFile = $"{RuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{RuntimeConfigLoader.CONFIG_EXTENSION}";
+            string cosmosFile = $"{FileSystemRuntimeConfigLoader.CONFIGFILE_NAME}.{COSMOS_ENVIRONMENT}{FileSystemRuntimeConfigLoader.CONFIG_EXTENSION}";
             RuntimeConfig overrides = new(
                 null,
                 new DataSource(DatabaseType.CosmosDB_NoSQL, $"AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==;Database={COSMOS_DATABASE_NAME}", new()),
@@ -1834,10 +1834,10 @@ namespace Azure.DataApiBuilder.Service.Tests.Configuration
         public static string GetConnectionStringFromEnvironmentConfig(string environment)
         {
             FileSystem fileSystem = new();
-            string sqlFile = new RuntimeConfigLoader(fileSystem).GetFileNameForEnvironment(environment, considerOverrides: true);
+            string sqlFile = new FileSystemRuntimeConfigLoader(fileSystem).GetFileNameForEnvironment(environment, considerOverrides: true);
             string configPayload = File.ReadAllText(sqlFile);
 
-            RuntimeConfigLoader.TryParseConfig(configPayload, out RuntimeConfig runtimeConfig);
+            FileSystemRuntimeConfigLoader.TryParseConfig(configPayload, out RuntimeConfig runtimeConfig);
 
             return runtimeConfig.DataSource.ConnectionString;
         }
