+++ src/Service.Tests/Unittests/MySqlQueryExecutorUnitTests.cs
@@ -52,8 +52,8 @@ namespace Azure.DataApiBuilder.Service.Tests.UnitTests
            );
 
             MockFileSystem fileSystem = new();
-            fileSystem.AddFile(RuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(mockConfig.ToJson()));
-            RuntimeConfigLoader loader = new(fileSystem);
+            fileSystem.AddFile(FileSystemRuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(mockConfig.ToJson()));
+            FileSystemRuntimeConfigLoader loader = new(fileSystem);
             RuntimeConfigProvider provider = new(loader);
             Mock<DbExceptionParser> dbExceptionParser = new(provider);
             Mock<ILogger<MySqlQueryExecutor>> queryExecutorLogger = new();
