+++ src/Service.Tests/Unittests/RestServiceUnitTests.cs
@@ -117,8 +117,8 @@ namespace Azure.DataApiBuilder.Service.Tests.UnitTests
            );
 
             MockFileSystem fileSystem = new();
-            fileSystem.AddFile(RuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(mockConfig.ToJson()));
-            RuntimeConfigLoader loader = new(fileSystem);
+            fileSystem.AddFile(FileSystemRuntimeConfigLoader.DEFAULT_CONFIG_FILE_NAME, new MockFileData(mockConfig.ToJson()));
+            FileSystemRuntimeConfigLoader loader = new(fileSystem);
             RuntimeConfigProvider provider = new(loader);
             MsSqlQueryBuilder queryBuilder = new();
             Mock<DbExceptionParser> dbExceptionParser = new(provider);
